Moved `PrisonerBasicInfo` interface directly under dto
Again, this is used by Prison API and Prisoner Search API
classes.
