+++ src/main/kotlin/uk/gov/justice/digital/hmpps/incentivesapi/dto/PrisonerBasicInfo.kt
@@ -1,0 +1,7 @@
+package uk.gov.justice.digital.hmpps.incentivesapi.dto
+
+interface PrisonerBasicInfo {
+  val bookingId: Long
+  val prisonerNumber: String
+  val prisonId: String
+}
