+++ src/main/kotlin/uk/gov/justice/digital/hmpps/incentivesapi/dto/prisonapi/PrisonAPI.kt
@@ -2,6 +2,7 @@ package uk.gov.justice.digital.hmpps.incentivesapi.dto.prisonapi
 
 import org.springframework.format.annotation.DateTimeFormat
 import uk.gov.justice.digital.hmpps.incentivesapi.dto.PrisonerAlert
+import uk.gov.justice.digital.hmpps.incentivesapi.dto.PrisonerBasicInfo
 import uk.gov.justice.digital.hmpps.incentivesapi.service.PrisonerInfoForNextReviewDate
 import java.time.LocalDate
 import java.time.LocalDateTime
@@ -31,12 +32,6 @@ data class BookingFromDatePair(
   @param:DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
   val fromDate: LocalDateTime,
 )
-
-interface PrisonerBasicInfo {
-  val bookingId: Long
-  val prisonerNumber: String
-  val prisonId: String
-}
 
 data class PrisonerInfo(
   override val bookingId: Long,
